feat: Update script to increment versions

+++ helpers/updateVersion.sh
@@ -5,10 +5,10 @@ newVersion=$(sed -nr "s/.*version.*\"([0-9]+\.[0-9]+\.[0-9]+)\".*/\1/p" package.
 
 # Update version in tizen-help script
 sed -r -i'.bak' -e "s/version=\"[0-9]+\.[0-9]+\.[0-9]+\"/version=\"$newVersion\"/g" tizen-help
+sed -r -i'.bak' -e "s/version=\"[0-9]+\.[0-9]+\.[0-9]+\"/version=\"$newVersion\"/g" tizen-help.ps1
 
 # Remove temporal file created by sed
-if [ -f tizen-help.bak ]; then
-  rm tizen-help.bak
-fi
+[ -f tizen-help.bak ] && rm tizen-help.bak
+[ -f tizen-help.ps1.bak ] && rm tizen-help.ps1.bak
 
 git add tizen-help
